Create context in welcome screen and refactored account screens
